mongodocstore: various changes
- Move URL handling to separate files.

- Clarify how LowercaseFields works and why we have it.

- Improve testing of LowercaseFields.
